# Performance tuning for ustring table