The main change here is to tune the default size and shard count for the ustring table. The average workstation has changed quite a bit since I last measured the performance of this code, and a simple tweak to these constants gives a nice speed boost on the `ustring_test` stress test.

On a machine with 100+ hardware threads, the new code stays efficient much longer. Previously the wedge test maxed out around 14% efficiency, and its close to 100% under the new code.

Measuring the cost of making 2M ustrings with a single thread is around 0.5s, while with all threads running at once (one will make progress while the others wait) it takes 4.0s (vs. 24.2s before this change).

On a more modest machine with 32 hardware threads, the tuned code is also faster (27% efficiency for the wedge test becomes 90%). Cost of making 2M strings with all threads at once goes from 6.0s down to 2.5s.

There are lots of different possible usage patterns and obviously `ustring_test` is not perfect, but I think this will be a win overall.

I did a bit of housecleaning on the file at the same time, switching to C++11 idioms in a few places.