@fpsunflower 2019-08-18
I think this one is also just down to Mac libheif issues and clang-format.

@lgritz is there a simple way to run the `clang-format` step locally?