@lgritz 2019-08-18
Yes!

    make clang-format