@fpsunflower 2019-08-19
`make clang-format` doesn't seem to be working on my (spi) machine, its finding the system `clang-format` binary instead of the one that lives alongside the compiler. It also is not able to pick up the `.clang-format` project file and therefore trying to reformat the whole project with LLVM style :(