@fpsunflower 2019-08-19
Running `clang-format` manually seems to have done the right thing ... I think this one should pass.