@lgritz 2019-08-19
Weird. I usually do it on my laptop, so maybe I need to try it on the work Linux and see what's going wrong.